添加玩家 弹幕 判定点 阴阳玉 添加imp动画

+++ src/game/bullet_pool.hpp
@@ -37,10 +37,12 @@ public:
         godot::RID canvas_item_rid;
         
         godot::Vector2 position;
+        /// @brief 每帧速度
         godot::Vector2 velocity;
         godot::Vector2 base_velocity; 
         float rotation = 0.0f;
-        float lifetime = 0.0f; 
+        /// @brief 存活帧数
+        int lifetime = 0; 
         bool active = false;
 
         BulletConfig config;
